// src/types/geometry.types.ts
/**
 * Represents a point in 2D space.
 *
 * @interface Point2D
 * @property {number} x The x-coordinate.
 * @property {number} y The y-coordinate.
 */
export interface Point2D {
  x: number;
  y: number;
}

/**
 * Represents a 2D size or dimensions with width and height.
 *
 * @interface Size2D
 * @property {number} width The width dimension.
 * @property {number} height The height dimension.
 */
export interface Size2D {
  width: number;
  height: number;
}